Organize utilities- 50%
