+++ src/py/kaleido/_utils/path_tools.py
@@ -4,6 +4,8 @@ import glob
 import re
 from pathlib import Path
 from typing import TYPE_CHECKING
+from urllib.parse import urlparse
+from urllib.request import url2pathname
 
 import logistro
 
@@ -56,3 +58,20 @@ def determine_path(
                 f"Cannot reach path {path.parent}. Are all directories created?",
             )
     return full_path
+
+
+def get_path(p: str | Path) -> Path:
+    if isinstance(p, Path):
+        return p
+    elif not isinstance(p, str):
+        raise TypeError("Path should be a string or `pathlib.Path` object.")
+
+    parsed = urlparse(str(p))
+
+    return Path(
+        url2pathname(parsed.path) if parsed.scheme.startswith("file") else p,
+    )
+
+
+def is_httpish(p: str) -> bool:
+    return urlparse(str(p)).scheme.startswith("http")
